Extract auth and schema CLI commands
